+++ src/main/java/com/bizcub/inventoryItemGroups/RawGroup.java
@@ -1,0 +1,11 @@
+package com.bizcub.inventoryItemGroups;
+
+import net.minecraft.world.item.ItemStack;
+
+import java.util.ArrayList;
+
+public class RawGroup {
+    public String name;
+    public boolean hasTranslation;
+    public ArrayList<ItemStack> items = new ArrayList<>();
+}